Generates heat maps

+++ nonogramSolve/Puzzle.h
@@ -47,7 +47,7 @@ public:
 	bool Puzzle::greedyValidityFront(bool****& dpValidityGrid, int i, int j, int c);
 	bool Puzzle::greedy();	//Solves puzzle one tomography element at a time.  Meh efficieny.  Returns if it finished (parallel only)
 
-	int* Puzzle::tomographyHeat(Tomography*& tomographyAxis);
+	unsigned long long* Puzzle::tomographyHeat(Tomography*& tomographyAxis);
 
 	void Puzzle::placeTrivial();
 	void Puzzle::likeHuman();
